requirements.txt
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cs50==9.3.4
Flask==3.0.0
Flask-Session==0.5.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
